fix some sign-conversion warnings

+++ fcd/callconv/params_registry.h
@@ -96,7 +96,7 @@ public:
 private:
 	CallingConvention* cc;
 	ContainerType values;
-	size_t returnBegin;
+	ptrdiff_t returnBegin;
 	Stage stage;
 	bool vararg;
 	
@@ -140,7 +140,7 @@ public:
 	size_t parameters_size() const
 	{
 		auto range = parameters();
-		return range.end() - range.begin();
+		return size_t(range.end() - range.begin());
 	}
 	
 	llvm::iterator_range<iterator> returns()
